[008-FEAT] : RankUpStandard
