+++ src/main/java/com/example/board/controller/SignUpController.java
@@ -3,7 +3,7 @@ package com.example.board.controller;
 import static com.example.board.domain.type.UserType.ADMIN;
 import static com.example.board.domain.type.UserType.USER;
 
-import com.example.board.domain.dto.UserDto;
+import com.example.board.domain.form.UserForm.SignUp;
 import com.example.board.service.SignUpService;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -22,14 +22,16 @@ public class SignUpController {
 
     // 관리자 회원가입
     @PostMapping("/signup/admin")
-    public ResponseEntity<String> adminSignUp(@RequestBody @Valid UserDto.SignUp signUp) {
+    public ResponseEntity<String> adminSignUp(@RequestBody @Valid SignUp signUp) {
+
         return ResponseEntity.ok(signUpService.signUp(signUp, ADMIN));
     }
 
 
     // 회원 회원가입
     @PostMapping("/signup/user")
-    public ResponseEntity<String> userSignUp(@RequestBody @Valid UserDto.SignUp signUp) {
+    public ResponseEntity<String> userSignUp(@RequestBody @Valid SignUp signUp) {
+
         return ResponseEntity.ok(signUpService.signUp(signUp, USER));
     }
 }
