+++ src/main/java/com/example/board/domain/dto/RankUpStandardDto.java
@@ -1,11 +1,6 @@
 package com.example.board.domain.dto;
 
-import com.example.board.domain.entity.RankUpStandard;
 import com.example.board.domain.type.RankType;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -26,54 +21,6 @@ public class RankUpStandardDto {
         private RankType rankName;
         private Long boardCount;
         private Long commentCount;
-        private boolean autoRankUpFlag;
-    }
-
-
-    @Getter
-    @Setter
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class AddRankUpStandard {
-
-        @NotNull(message = "등급을 확인해주세요.")
-        @Enumerated(EnumType.STRING)
-        private RankType rankName;
-        @NotNull(message = "게시글 작성 수를 입력하세요.")
-        @Min(value = 0, message = "게시글 작성 수는 0이상 입니다.")
-        private Long boardCount;
-        @NotNull(message = "댓글 작성 수를 입력하세요.")
-        @Min(value = 0, message = "댓글 작성 수는 0이상 입니다.")
-        private Long commentCount;
-        @NotNull(message = "자동 등업 여부를 선택하세요.")
-        private boolean autoRankUpFlag;
-
-        public RankUpStandard save(AddRankUpStandard addRankUpStandard) {
-            return RankUpStandard.builder()
-                .rankName(addRankUpStandard.getRankName())
-                .boardCount(addRankUpStandard.getBoardCount())
-                .commentCount(addRankUpStandard.getCommentCount())
-                .autoRankUpFlag(addRankUpStandard.isAutoRankUpFlag())
-                .build();
-        }
-    }
-
-
-    @Getter
-    @Setter
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class ModifyRankUpStandard {
-
-        @NotNull(message = "게시글 작성 수를 입력하세요.")
-        @Min(value = 0, message = "게시글 작성 수는 0이상 입니다.")
-        private Long boardCount;
-        @NotNull(message = "댓글 작성 수를 입력하세요.")
-        @Min(value = 0, message = "댓글 작성 수는 0이상 입니다.")
-        private Long commentCount;
-        @NotNull(message = "자동 등업 여부를 선택하세요.")
         private boolean autoRankUpFlag;
     }
 }
