+++ src/main/java/com/example/board/domain/form/UserForm.java
@@ -1,8 +1,5 @@
-package com.example.board.domain.dto;
+package com.example.board.domain.form;
 
-import com.example.board.domain.entity.User;
-import com.example.board.domain.type.RankType;
-import com.example.board.domain.type.UserType;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
@@ -11,8 +8,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
-public class UserDto {
+public class UserForm {
 
     @Getter
     @Setter
@@ -28,17 +24,8 @@ public class UserDto {
         private String userNickName;
         @NotBlank(message = "비밀번호를 입력하세요.")
         @Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{10,}",
-                 message = "비밀번호는 10자 이상, 영문 대 소문자, 숫자, 특수문자를 사용하세요.")
+            message = "비밀번호는 10자 이상, 영문 대 소문자, 숫자, 특수문자를 사용하세요.")
         private String userPassword;
-
-        public User save(SignUp signUp, UserType userType) {
-            return User.builder()
-                .loginId(signUp.getLoginId())
-                .userNickName(signUp.getUserNickName())
-                .userPassword(signUp.getUserPassword())
-                .userRank(userType.equals(UserType.USER) ? RankType.LEVEL1 : RankType.ADMIN)
-                .build();
-        }
     }
