+++ src/main/java/com/example/board/service/SignInService.java
@@ -1,9 +1,9 @@
 package com.example.board.service;
 
-import com.example.board.domain.dto.UserDto;
+import com.example.board.domain.form.UserForm.SignIn;
 
 public interface SignInService {
 
     // 로그인
-    String signIn(UserDto.SignIn signIn);
+    String signIn(SignIn signIn);
 }
