+++ src/main/java/com/example/board/service/SignUpService.java
@@ -1,11 +1,11 @@
 package com.example.board.service;
 
-import com.example.board.domain.dto.UserDto;
+import com.example.board.domain.form.UserForm.SignUp;
 import com.example.board.domain.type.UserType;
 
 
 public interface SignUpService {
 
     // 회원가입
-    String signUp(UserDto.SignUp signUp, UserType signLocation);
+    String signUp(SignUp signUp, UserType signLocation);
 }
