+++ src/main/java/com/example/board/service/impl/SignInServiceImpl.java
@@ -3,8 +3,8 @@ package com.example.board.service.impl;
 import static com.example.board.exception.ErrorCode.NOT_FIND_LOGIN_ID;
 import static com.example.board.exception.ErrorCode.WRONG_LOGIN_PASSWORD;
 
-import com.example.board.domain.dto.UserDto;
 import com.example.board.domain.entity.User;
+import com.example.board.domain.form.UserForm.SignIn;
 import com.example.board.domain.repository.UserRepository;
 import com.example.board.exception.GlobalException;
 import com.example.board.security.TokenProvider;
@@ -24,7 +24,8 @@ public class SignInServiceImpl implements SignInService {
 
 
     // 로그인
-    public String signIn(UserDto.SignIn signIn) {
+    public String signIn(SignIn signIn) {
+
         User user = userRepository.findByLoginId(signIn.getLoginId())
             .orElseThrow(() -> new GlobalException(NOT_FIND_LOGIN_ID));
