+++ src/main/java/com/example/board/service/impl/SignUpServiceImpl.java
@@ -3,8 +3,10 @@ package com.example.board.service.impl;
 import static com.example.board.exception.ErrorCode.ALREADY_LOGIN_ID;
 import static com.example.board.exception.ErrorCode.ALREADY_NICK_NAME;
 
-import com.example.board.domain.dto.UserDto;
+import com.example.board.domain.entity.User;
+import com.example.board.domain.form.UserForm.SignUp;
 import com.example.board.domain.repository.UserRepository;
+import com.example.board.domain.type.RankType;
 import com.example.board.domain.type.UserType;
 import com.example.board.exception.GlobalException;
 import com.example.board.service.SignUpService;
@@ -24,7 +26,8 @@ public class SignUpServiceImpl implements SignUpService {
 
     // 회원가입
     @Transactional
-    public String signUp(UserDto.SignUp signUp, UserType userType) {
+    public String signUp(SignUp signUp, UserType userType) {
+
         // 아이디 확인
         if (userRepository.existsByLoginId(signUp.getLoginId())) {
             throw new GlobalException(ALREADY_LOGIN_ID);
@@ -35,8 +38,13 @@ public class SignUpServiceImpl implements SignUpService {
             throw new GlobalException(ALREADY_NICK_NAME);
         }
 
-        signUp.setUserPassword(passwordEncoder.encode(signUp.getUserPassword()));
-        userRepository.save(signUp.save(signUp, userType));
+        userRepository.save(
+            User.builder()
+                .loginId(signUp.getLoginId())
+                .userNickName(signUp.getUserNickName())
+                .userPassword(passwordEncoder.encode(signUp.getUserPassword()))
+                .userRank(userType.equals(UserType.USER) ? RankType.LEVEL1 : RankType.ADMIN)
+                .build());
 
         return "회원가입이 완료 되었습니다.";
     }
